Add player turn prompt messages
I added messages, which show up on
each player's respective sides, to
prompt them that it is their turn to
move. The way I have it organized,
min width of the body is now 1200px,
which may be too wide, though I may
be able to make do with media query
to handle any screen < 1200. I think,
though, that ideal media-query is
around 800, not 1200. So I'll see how
it works. The mac may not be the best
place to test, because it has higher
res than most computers.