Add general game/win logic to scripts.js
I wrote javascript to establish
end-game conditions, stopping the game
and declaring a winner if those
conditions are met (i.e. either three
Xs or three Os in a row), or declaring
a draw. My code works but needs desperately
to be refactored. And this version currently
only supports a single match.